+++ src/views/components/SearchFilter/SearchFilter.tsx
@@ -1,6 +1,6 @@
 import { ISearchFilter } from './ISearchFilter'
 import style from './SearchFilter.module.scss'
-import { MetadataFilters } from '@views/components'
+import { FilterSort, MetadataFilters } from '@views/components'
 import {
 	SearchFilterViewModel
 } from '@src/view-models/SearchFilterViewModel.ts'
@@ -8,9 +8,6 @@ import { useContext, useState } from 'react'
 import { DataRequestParams } from '@data/request-parameters'
 import { useSearchPage } from '@src/hooks/useSearchPage.ts'
 import { RootContext } from '@data/context'
-import {
-	ISearchFilterViewModel
-} from '@src/view-models/interfaces/ISearchFilterViewModel.ts'
 
 
 export { SearchFilter }
@@ -35,26 +32,3 @@ function SearchFilter(props: ISearchFilter) {
 	)
 }
 
-interface IFilterSort {
-	parentViewModel: ISearchFilterViewModel
-}
-
-function FilterSort({ parentViewModel }: IFilterSort) {
-	const orderingOptions = parentViewModel.ordering.map(o => {
-		return (
-			<option key={o}
-					value={o}>
-				{o}
-			</option>
-		)
-	})
-
-	return (
-		<p>
-			<label htmlFor="ordering">order by</label>
-			<select name="result_order" id="ordering">
-				{orderingOptions}
-			</select>
-		</p>
-	)
-}
