+++ src/views/components/index.ts
@@ -29,3 +29,4 @@ export { MainMenu } from '@views/components/Header'
 export { CardScore } from '@views/components/CardScore/CardScore.tsx'
 export { CardHeader } from '@views/components/CardHeader/CardHeader.tsx'
 export { CardMeta } from '@views/components/CardMeta/CardMeta.tsx'
+export { FilterSort } from '@views/components/FilterSort/FilterSort.tsx'
